+++ index.html
@@ -226,7 +226,6 @@
       </div>
     </section>
   </template>
-<script src="functions.js"></script>
-<script src="main.js"></script>
+<script src="main.js" type="module"></script>
 </body>
 </html>
